+++ tango/__init__.py
@@ -12,9 +12,12 @@ __all__ = [
     "Step",
     "StepCache",
     "LocalStepCache",
+    "Workspace",
+    "MemoryWorkspace",
+    "LocalWorkspace",
 ]
 
-from tango.executor import Executor, SimpleExecutor
+from tango.executor import Executor
 from tango.format import (
     DillFormat,
     DillFormatIterator,
@@ -22,6 +25,8 @@ from tango.format import (
     JsonFormat,
     JsonFormatIterator,
 )
+from tango.local_workspace import LocalWorkspace
 from tango.step import Step
 from tango.step_cache import LocalStepCache, StepCache
-from tango.step_graph import StepGraph, StepStub
+from tango.step_graph import StepGraph
+from tango.workspace import MemoryWorkspace, Workspace