+++ tango/common/file_lock.py
@@ -4,7 +4,6 @@ import warnings
 from filelock import AcquireReturnProxy
 from filelock import FileLock as _FileLock
 from filelock import Timeout
-from overrides import overrides
 
 from .aliases import PathOrStr
 
@@ -24,10 +23,9 @@ class FileLock(_FileLock):  # type: ignore[valid-type,misc]
         super().__init__(str(lock_file), timeout=timeout)
         self._read_only_ok = read_only_ok
 
-    @overrides
-    def acquire(self, timeout=None, poll_intervall=0.05) -> AcquireReturnProxy:
+    def acquire(self, timeout=None, poll_interval=0.05) -> AcquireReturnProxy:
         try:
-            return super().acquire(timeout=timeout, poll_intervall=poll_intervall)
+            return super().acquire(timeout=timeout, poll_interval=poll_interval)
         except OSError as err:
             # OSError could be a lot of different things, but what we're looking
             # for in particular are permission errors, such as: