+++ tango/common/testing.py
@@ -3,12 +3,11 @@ import os
 import shutil
 import tempfile
 from contextlib import contextmanager
-from copy import deepcopy
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Union, cast
 
-from .aliases import PathOrStr
-from .registrable import Registrable
+from tango.common.aliases import PathOrStr
+from tango.common.logging import initialize_logging
 
 
 class TangoTestCase:
@@ -54,19 +53,8 @@ class TangoTestCase:
         self.TEST_DIR = Path(tempfile.mkdtemp(prefix="tango_tests"))
         os.makedirs(self.TEST_DIR, exist_ok=True)
 
-    @classmethod
-    def setup_class(cls):
-        # During teardown we'll restore the state of `Registrable`'s internal registry
-        # to make sure any registered mock test classes are removed so they don't conflict
-        # with other tests.
-        cls._original_registry = deepcopy(Registrable._registry)
-
     def teardown_method(self):
         shutil.rmtree(self.TEST_DIR)
-
-    @classmethod
-    def teardown_class(cls):
-        Registrable._registry = cls._original_registry
 
     def run(
         self,
@@ -88,15 +76,13 @@ class TangoTestCase:
 
             overrides = json.dumps(overrides)
 
-        run_dir = self.TEST_DIR / "run"
-        _run(
+        return _run(
             TangoGlobalSettings(),
             str(config),
-            directory=str(run_dir),
+            workspace_dir=str(self.TEST_DIR / "workspace"),
             overrides=overrides,
             include_package=include_package,
         )
-        return run_dir
 
 
 @contextmanager
@@ -108,6 +94,7 @@ def run_experiment(
     the experiment and returns the path to the cache directory, a temporary directory that will be
     cleaned up on ``__exit__``.
     """
+    initialize_logging(enable_click_logs=True)
     test_case = TangoTestCase()
     try:
         test_case.setup_method()